refactor(frontend): move component to correct module
