+++ apps/frontend/app/routes/_dashboard.media.groups.$id._index.tsx
@@ -27,9 +27,11 @@ import { useState } from "react";
 import invariant from "tiny-invariant";
 import { z } from "zod";
 import { zx } from "zodix";
-import { MediaDetailsLayout } from "~/components/common";
 import {
 	AddEntityToCollectionModal,
+	MediaDetailsLayout,
+} from "~/components/common";
+import {
 	DisplayCollection,
 	MediaScrollArea,
 	PartialMetadataDisplay,
