+++ apps/frontend/app/routes/_dashboard.media.item.$id._index.tsx
@@ -94,9 +94,11 @@ import invariant from "tiny-invariant";
 import { match } from "ts-pattern";
 import { z } from "zod";
 import { zx } from "zodix";
-import { MediaDetailsLayout } from "~/components/common";
 import {
 	AddEntityToCollectionModal,
+	MediaDetailsLayout,
+} from "~/components/common";
+import {
 	DisplayCollection,
 	MediaScrollArea,
 	PartialMetadataDisplay,
